+++ src/async_monitored_item.rs
@@ -543,6 +543,11 @@ macro_rules! data_change_impl {
     };
 }
 
+// Note: Array values are not supported yet in their typed form: previously, any such attempt would
+// fail, because converting to `DataValue` expects scalar values.
+//
+// To give us some time to think about the best, typed representation of such non-scalar values, we
+// remove their `impl` for now. Access is still possible with the non-typed attribute methods.
 data_change_impl!(
     NodeId,
     NodeClass,
@@ -550,6 +555,7 @@ data_change_impl!(
     DisplayName,
     Description,
     WriteMask,
+    UserWriteMask,
     IsAbstract,
     Symmetric,
     InverseName,
@@ -559,12 +565,18 @@ data_change_impl!(
     Value,
     DataType,
     ValueRank,
-    ArrayDimensions,
+    // ArrayDimensions,
     AccessLevel,
-    AccessLevelEx,
+    UserAccessLevel,
     MinimumSamplingInterval,
     Historizing,
     Executable,
+    UserExecutable,
+    DataTypeDefinition,
+    // RolePermissions,
+    // UserRolePermissions,
+    AccessRestrictions,
+    AccessLevelEx,
 );
 
 impl MonitoredItemAttribute for attributes::EventNotifier {